Run tests against `sync` and `unsync` equally.
This does not benefit the individual tests a lot, but it should help
catch any unintended divergence between what the two “flavors” offer.

I also moved other tests not being split in this fashion into
/tests/api because it is good practice to test a library from outside
rather than inside, to prove that private items are not accidentally
being used.
