+++ src/notifier.rs
@@ -394,72 +394,4 @@ impl<M, L> Clone for NotifierForwarder<M, L> {
 // -------------------------------------------------------------------------------------------------
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::Sink;
-    use alloc::sync::Arc;
-    use alloc::{format, vec};
-
-    #[test]
-    fn notifier_basics_and_debug() {
-        let cn: crate::unsync::Notifier<u8> = Notifier::new();
-        assert_eq!(format!("{cn:?}"), "Notifier(0)");
-        cn.notify(&0);
-        assert_eq!(format!("{cn:?}"), "Notifier(0)");
-        let sink = Sink::new();
-        cn.listen(sink.listener());
-        assert_eq!(format!("{cn:?}"), "Notifier(1)");
-        // type annotation to prevent spurious inference failures in the presence
-        // of other compiler errors
-        assert_eq!(sink.drain(), Vec::<u8>::new());
-        cn.notify(&1);
-        cn.notify(&2);
-        assert_eq!(sink.drain(), vec![1, 2]);
-        assert_eq!(format!("{cn:?}"), "Notifier(1)");
-    }
-
-    // Test for NotifierForwarder functionality exists as a doc-test.
-
-    #[test]
-    fn notifier_forwarder_debug() {
-        let n: Arc<crate::unsync::Notifier<u8>> = Arc::new(Notifier::new());
-        let nf = Notifier::forwarder(Arc::downgrade(&n));
-        assert_eq!(
-            format!("{nf:?}"),
-            "NotifierForwarder { alive(shallow): true, .. }"
-        );
-        drop(n);
-        assert_eq!(
-            format!("{nf:?}"),
-            "NotifierForwarder { alive(shallow): false, .. }"
-        );
-    }
-
-    #[test]
-    fn close_drops_listeners() {
-        #[derive(Debug)]
-        struct DropDetector(Arc<()>);
-        impl Listener<()> for DropDetector {
-            fn receive(&self, _: &[()]) -> bool {
-                true
-            }
-        }
-
-        let notifier = crate::unsync::Notifier::<()>::new();
-        let detector = DropDetector(Arc::new(()));
-        let weak = Arc::downgrade(&detector.0);
-        notifier.listen(detector);
-
-        assert_eq!(weak.strong_count(), 1);
-        notifier.close();
-        assert_eq!(weak.strong_count(), 0);
-    }
-
-    #[test]
-    #[should_panic = "cannot send messages after Notifier::close()"]
-    fn notify_after_close() {
-        let notifier = crate::unsync::Notifier::<()>::new();
-        notifier.close();
-        notifier.notify(&());
-    }
-}
+mod tests {}
