+++ src/source.rs
@@ -226,42 +226,3 @@ where
         Arc::new(value)
     }
 }
-
-// -------------------------------------------------------------------------------------------------
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::unsync;
-    use crate::Sink;
-    use alloc::format;
-    use pretty_assertions::assert_eq;
-
-    #[test]
-    fn constant_source_debug() {
-        let source = unsync::constant(123);
-        assert_eq!(format!("{source:?}"), "Constant(123)");
-        assert_eq!(format!("{source:#?}"), "Constant(\n    123,\n)");
-    }
-
-    #[test]
-    fn constant_source_usage() {
-        let sink: Sink<()> = Sink::new();
-        let source = unsync::constant(123u64);
-        source.listen(sink.listener()); // doesn't panic, doesn't do anything
-        assert_eq!(source.get(), 123);
-        assert_eq!(source.get(), 123);
-        assert!(sink.drain().is_empty());
-    }
-
-    #[cfg(feature = "sync")] // because `Sink` isn't `Sync` otherwise
-    #[test]
-    fn constant_source_usage_sync() {
-        let sink: Sink<()> = Sink::new();
-        let source = crate::sync::constant(123u64);
-        source.listen(sink.listener()); // doesn't panic, doesn't do anything
-        assert_eq!(source.get(), 123);
-        assert_eq!(source.get(), 123);
-        assert!(sink.drain().is_empty());
-    }
-}
